DKCC

DKCC

DKCC(seurat, threshold = 0.7, max.iter = 1)

Arguments

seurat
seurat object

threshold
minimum value for an identity to be assigned within the model call, default is 0.7

max.iter
Can ask scPred to run this number of integrations, set to 0 be default

Value

seurat object with additional metadata columns

Examples

organoid <- DKCC(organoid)

#> ●  Matching reference with new dataset...
#> 	 ─ 9977 features present in reference loadings
#> 	 ─ 7520 features shared between reference and new dataset
#> 	 ─ 75.37% of features in the reference are present in new dataset
#> ●  Aligning new data to reference...
#> 

#> Harmony 1/1

#> ●  Classifying cells...
#> DONE!
#> ‒ Data has already being aligned to a reference.
#> ⁺ Skip data alignment using `recompute.alignment = FALSE`.
#> ●  Matching reference with new dataset...
#> 	 ─ 9768 features present in reference loadings
#> 	 ─ 7427 features shared between reference and new dataset
#> 	 ─ 76.03% of features in the reference are present in new dataset
#> ●  Aligning new data to reference...
#> 

#> Harmony 1/1

#> ●  Classifying cells...
#> DONE!
#> ‒ Data has already being aligned to a reference.
#> ⁺ Skip data alignment using `recompute.alignment = FALSE`.
#> ●  Matching reference with new dataset...
#> 	 ─ 8140 features present in reference loadings
#> 	 ─ 6605 features shared between reference and new dataset
#> 	 ─ 81.14% of features in the reference are present in new dataset
#> ●  Aligning new data to reference...
#> 

#> Harmony 1/1

#> ●  Classifying cells...
#> DONE!
#> ‒ Data has already being aligned to a reference.
#> ⁺ Skip data alignment using `recompute.alignment = FALSE`.
#> ●  Matching reference with new dataset...
#> 	 ─ 8693 features present in reference loadings
#> 	 ─ 7037 features shared between reference and new dataset
#> 	 ─ 80.95% of features in the reference are present in new dataset
#> ●  Aligning new data to reference...
#> 

#> Harmony 1/1

#> ●  Classifying cells...
#> DONE!
#> ‒ Data has already being aligned to a reference.
#> ⁺ Skip data alignment using `recompute.alignment = FALSE`.
#> ●  Matching reference with new dataset...
#> 	 ─ 8794 features present in reference loadings
#> 	 ─ 6956 features shared between reference and new dataset
#> 	 ─ 79.1% of features in the reference are present in new dataset
#> ●  Aligning new data to reference...
#> 

#> Harmony 1/1

#> ●  Classifying cells...
#> DONE!
#> ‒ Data has already being aligned to a reference.
#> ⁺ Skip data alignment using `recompute.alignment = FALSE`.
#> ●  Matching reference with new dataset...
#> 	 ─ 8736 features present in reference loadings
#> 	 ─ 6820 features shared between reference and new dataset
#> 	 ─ 78.07% of features in the reference are present in new dataset
#> ●  Aligning new data to reference...
#> 

#> Harmony 1/1

#> ●  Classifying cells...
#> DONE!
#> 

#> Warning: pseudoinverse used at -1.4983

#> Warning: neighborhood radius 0.30103

#> Warning: reciprocal condition number  9.4429e-15

#> Centering and scaling data matrix

#> PC_ 1 
#> Positive:  NEB, MYL1, VGLL2, MYL4, PDLIM3, RPS18, RPL12, SGCA, CDH15, RPL7A 
#> 	   RPS12, NEXN, RPS17, NRK, MYOD1, ACTC1, RPS25, USP18, MYLPF, RPS2 
#> 	   RP1-302G2.5, EDN3, PALMD, ARPP21, FGFR4, RPL29, RAPSN, MYOG, MRLN, CKB 
#> Negative:  MDK, ITM2C, NR2F2, CRABP2, IFITM3, FTL, TPM1, S100A13, CD24, HOXA10 
#> 	   NR2F1, HMGA2, COL2A1, IFITM2, LYPD1, WT1, BST2, PAX8, HMGN3, TSHZ2 
#> 	   CTSB, APOE, ZNF503, SEPT6, PAX2, HOXC10, CCDC80, CD59, HSP90AA1, OCIAD2 
#> PC_ 2 
#> Positive:  TOP2A, CENPF, MELK, CCNB2, CENPN, MKI67, CDK1, UBE2C, MAD2L1, SLC2A8 
#> 	   KIFC1, PHF19, KIF23, ASPM, CDCA3, NUSAP1, HMGB2, NCAPG, HJURP, SPC25 
#> 	   POLE2, SMC4, CENPM, CENPU, CENPW, H2AFZ, ARHGAP11B, HIST1H4C, RAD51AP1, PIF1 
#> Negative:  SELM, KLHDC8B, GAP43, GDAP1, OAF, EBF1, VIM, COL5A2, LGALS1, CADM3 
#> 	   AFF3, PDGFRA, COL1A2, TMSB4X, ISL1, SOX2, SOX4, AGO1, GNG11, ZIC2 
#> 	   MAPK8IP1, KCNQ2, SHISA9, ZIC5, RAPGEF5, DNM3OS, PAX3, THSD4, CDH11, DDIT4 
#> PC_ 3 
#> Positive:  COL1A2, COL3A1, ZFHX4, ITM2A, LDHA, DNM3OS, NFIB, RSL1D1, MSC, STC1 
#> 	   SPARCL1, ENO1, COL1A1, RPL22L1, MAGED2, MDFI, EMP3, DCN, MAGED1, MYF5 
#> 	   OAF, EIF4EBP1, NCL, VIM, CALD1, C1QTNF3, KCNE5, HSP90AA1, PDGFRA, SPG20 
#> Negative:  ACTA1, TAGLN3, SST, RAB26, LACE1, ONECUT2, OLFM1, ASS1, SHISA9, MARCH9 
#> 	   KLHL25, KCNQ2, CADM3, ERICH3, ELAVL4, HOTAIR, NTM, PCDH19, FAM110A, FAM57B 
#> 	   SSTR2, FOXD3-AS1, PPP1R17, WFIKKN1, ZIC5, RAPGEF5, FYCO1, KIAA1456, DCX, RGS9 
#> PC_ 4 
#> Positive:  ASPM, HIST1H4C, KIAA0101, SKA3, RGS19, TK1, H2AFX, KIF20B, CDK1, SGOL2 
#> 	   AURKB, ASF1B, NDC80, ELAVL2, FBXO5, CDKN3, FAM111A, ZWINT, CKS1B, MDH1B 
#> 	   COL1A2, SMC4, CCNA2, HOXA7, DUT, RP1-79C4.4, GAS2, ST3GAL3, UBE2C, PBK 
#> Negative:  GLIPR2, BST2, BCAM, PVRL2, PAX2, WT1, DSP, CLU, FRMD4B, CITED2 
#> 	   EPCAM, TRIM37, KRT19, PERP, PAX8, CLTA, KRT18, CTTNBP2, TBXAS1, PON2 
#> 	   LBH, YPEL5, LYPD1, IGFBP7, NR2F1, HIPK2, ADCY2, KRT8, TRABD2B, C8orf4 
#> PC_ 5 
#> Positive:  COBL, SYNPO2L, HFE2, APOBEC2, TTN, NPNT, WNT10A, AC018647.3, SCRIB, CTD-2545M3.8 
#> 	   SNTB1, FILIP1, WNT9A, SEPT4, KLHL41, ITGA4, LRRN1, STAC3, MYOG, TNFSF13B 
#> 	   AC016700.5, FNDC5, TMCC2, MYH3, SIRT2, LMOD3, TNNC2, RASSF4, RXRG, LMNA 
#> Negative:  MSC, C1QTNF3, MYF5, ITM2A, FABP5, AKR1C1, MATN2, NELL2, CHODL, SSTR2 
#> 	   DCN, SPG20, CRMP1, STC1, MYC, PAX7, REEP1, MARCH9, CCT5, DKC1 
#> 	   RGCC, NTF3, GPRIN3, TAGLN2, NDST3, ADA, YBX1, PITX2, CDC37L1, SST 

#> Warning: The default method for RunUMAP has changed from calling Python UMAP via reticulate to the R-native UWOT using the cosine metric
#> To use Python UMAP via reticulate, set umap.method to 'umap-learn' and metric to 'correlation'
#> This message will be shown once per session

#> 16:49:38 UMAP embedding parameters a = 0.9922 b = 1.112

#> 16:49:38 Read 63 rows and found 20 numeric columns

#> 16:49:38 Using Annoy for neighbor search, n_neighbors = 30

#> 16:49:38 Building Annoy index with metric = cosine, n_trees = 50

#> 0%   10   20   30   40   50   60   70   80   90   100%

#> [----|----|----|----|----|----|----|----|----|----|

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> *

#> |

#> 16:49:38 Writing NN index file to temp file /tmp/RtmphHEAlw/file6f5207610b8

#> 16:49:38 Searching Annoy index using 1 thread, search_k = 3000

#> 16:49:38 Annoy recall = 100%

#> 16:49:39 Commencing smooth kNN distance calibration using 1 thread

#> 16:49:39 Initializing from normalized Laplacian + noise

#> 16:49:39 Commencing optimization for 500 epochs, with 1924 positive edges

#> 16:49:40 Optimization finished

#> Computing nearest neighbor graph

#> Computing SNN

#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 63
#> Number of edges: 1823
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.5100
#> Number of communities: 2
#> Elapsed time: 0 seconds

#> Warning: Only less than three identities present, the expression values will be not scaled