DotPlotCompare

DotPlotCompare

DotPlotCompare(
  new.object = NULL,
  show = "gene",
  split.by = "orig.ident",
  scaling = "log",
  size = "pct.exp",
  idents = "all",
  classification = "DKCC",
  dot.min = 0.1,
  dot.scale = 6,
  col.min = NA,
  col.max = NA,
  scale.min = NA,
  scale.max = NA,
  scale.by = "radius",
  features,
  compare.to.organoids = FALSE,
  filter.samples = NULL,
  columns = NULL
)

Arguments

new.object
seurat object to compare to reference data

show
this defines the focus of the output; "gene" partitions by gene, anything else partitions by segment

split.by
meta.data column to split the sample by, usually the sample column

scaling
choose "log", "scale" or "raw" to change how the expression values are displayed

size
how to scale the size of the dot, "pct.exp" for the percent of cells expressing the gene, "Pct" is percent of cells in identity

idents
Identities to display. Use "all" to show all segments, "nephron" to show all nephron segments, "others" to show all non-nephron segments, or a string of identities to select specific Identities

classification
choose "DKCC" or "LineageID"

dot.min
dot size of minimum cell percentage

dot.scale
dot size of maximum cell percentage

col.min
minimum expression value to plot

col.max
maximum expression value to plot

scale.min
minimum scale

scale.max
maximum scale

scale.by
"radius" or "size"

features
features to plot

compare.to.organoids
if function should plot organoid samples from the database

filter.samples
if compare.to.organoids=T, can provide desired sample names here using the output from "GetSampleIDs".

columns
how many columns to plot. Default is to plot as square as possible.

Value

ggplot object